Extracting legacy P2P protocol into lnp2p crate
