+++ Cargo.toml
@@ -31,7 +31,6 @@ secp256k1 = { version = "0.20.3", features = ["global-context-less-secure"] }
 internet2 = "0.5.0-alpha.2"
 serde_crate = { package = "serde", version = "1", features = ["derive"], optional = true }
 serde_with = { version = "1.8", features = ["hex"], optional = true }
-lazy_static = "1.4"
 
 [features]
 default = []
