+++ lnp2p/Cargo.toml
@@ -18,14 +18,17 @@ path = "src/lib.rs"
 crate-type = ["rlib", "staticlib"]
 
 [dependencies]
-amplify = "3.9.1"
+amplify = { version = "3.9.1", features = ["hex"] }
+lnpbp = "0.5.0-beta.3"
 rgb-core = { version = "0.5.0-alpha.1", optional = true }
 strict_encoding = { version = "1.7.4", optional = true }
 lightning_encoding = { version = "0.5.0-beta.3", optional = true }
 descriptor-wallet = "0.5.0-alpha.5"
 bitcoin = "0.27.1"
+internet2 = "0.5.0-alpha.2"
 serde_crate = { package = "serde", version = "1", features = ["derive"], optional = true }
 serde_with = { version = "1.5", features = ["hex"], optional = true }
+lazy_static = "1.4"
 
 [features]
 default = ["legacy"]
